requirements.txt
django
selenium
simplejson
sphinx
django-simple-history